pyproject.toml
[tool.poetry]
name = "cdxjl"
version = "0.1.0"
description = ""
authors = ["Richard Estwick <richard@cybeats,com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.2"
cyclonedx-python-lib = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
